fix warning about non-final resource IDs

+++ src/main/java/org/thoughtcrime/securesms/ContactSelectionListFragment.java
@@ -149,16 +149,16 @@ public class ContactSelectionListFragment extends    Fragment
 
       @Override
       public boolean onActionItemClicked(ActionMode actionMode, MenuItem menuItem) {
-        switch (menuItem.getItemId()) {
-          case R.id.menu_select_all:
-            handleSelectAll();
-            return true;
-          case R.id.menu_view_profile:
-            handleViewProfile();
-            return true;
-          case R.id.menu_delete_selected:
-            handleDeleteSelected();
-            return true;
+        int itemId = menuItem.getItemId();
+        if (itemId == R.id.menu_select_all) {
+          handleSelectAll();
+          return true;
+        } else if (itemId == R.id.menu_view_profile) {
+          handleViewProfile();
+          return true;
+        } else if (itemId == R.id.menu_delete_selected) {
+          handleDeleteSelected();
+          return true;
         }
         return false;
       }
